+++ braintree/transparent_redirect_gateway.py
@@ -1,17 +1,11 @@
 import cgi
-import sys
 from datetime import datetime
-if sys.version_info[0] == 2:
-    from urllib import urlencode
-else:
-    from urllib.parse import urlencode
-import braintree
+from six.moves.urllib.parse import urlencode
 from braintree.util.crypto import Crypto
-from braintree.error_result import ErrorResult
 from braintree.exceptions.forged_query_string_error import ForgedQueryStringError
 from braintree.util.http import Http
-from braintree.successful_result import SuccessfulResult
 from braintree.transparent_redirect import TransparentRedirect
+
 
 class TransparentRedirectGateway(object):
     def __init__(self, gateway):
@@ -58,7 +52,7 @@ class TransparentRedirectGateway(object):
         query_params = cgi.parse_qs(query_string)
         http_status = int(query_params["http_status"][0])
         message = query_params.get("bt_message")
-        if message != None:
+        if message is not None:
             message = message[0]
 
         if Http.is_error_status(http_status):