+++ braintree/util/generator.py
@@ -1,15 +1,7 @@
 import datetime
-import sys
+import six
 from decimal import Decimal
 
-if sys.version_info[0] == 2:
-    integer_types = int, long
-    text_type = unicode
-    binary_type = str
-else:
-    integer_types = int,
-    text_type = str
-    binary_type = bytes
 
 class Generator(object):
     def __init__(self, dict):
@@ -43,9 +35,9 @@ class Generator(object):
         open_tag = "<" + self.__escape(key) + ">"
         close_tag = "</" + self.__escape(key) + ">"
 
-        if isinstance(value, text_type):
+        if isinstance(value, six.string_types):
             return open_tag + self.__escape(value).encode('ascii', 'xmlcharrefreplace').decode('utf-8') + close_tag
-        elif isinstance(value, binary_type):
+        elif isinstance(value, six.binary_type):
             return open_tag + self.__escape(value) + close_tag
         elif isinstance(value, Decimal):
             return open_tag + str(value) + close_tag
@@ -57,7 +49,7 @@ class Generator(object):
         elif isinstance(value, bool):
             open_tag = "<" + key + " type=\"boolean\">"
             return open_tag + self.__generate_boolean(value) + close_tag
-        elif isinstance(value, integer_types) and not isinstance(value, bool):
+        elif isinstance(value, six.integer_types) and not isinstance(value, bool):
             open_tag = "<" + key + " type=\"integer\">"
             return open_tag + str(value) + close_tag
         elif isinstance(value, type(None)):